added project files for (new file structure):  - eVC4  - VC6  - VS.NET  - VC8 (desktop)  - VC8 (WinCE) fixed several files to compile with VC8.CE

+++ src/common/CommonNodeFactory.cxx
@@ -463,7 +463,7 @@ Node * CommonNodeFactory::createNode( const std::string& name, StringTable& attr
         if( attributes.get("min", &min) != 1 )
             min = 0;
         if( attributes.get("max", &max ) != 1 )
-            max = sqrt(FLT_MAX)-0.1f;
+            max = (float)sqrt(FLT_MAX)-0.1f;
         result = new RangeFilterNode( min, max );
     }
     else if( name.compare("PositionFilterNode") == 0 )
